+++ my-app/src/components/student/styles/students.module.css
@@ -1,0 +1,25 @@
+.card {
+    width: 15em;
+    margin: 0.5em;
+}
+
+.students {
+    text-align: center;
+    width: 100%;
+    height: 100vh;
+}
+
+.formStudent {
+    padding: 1em;
+    text-align: left;
+}
+
+.marginR1 {
+    margin-right: 1em;
+}
+
+.inputStudent {
+    background-color: transparent;
+    border: 0;
+    border-bottom: 1px solid white;
+}
