+++ my-app/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import './styles/index.css';
+import App from './components/App';
 import * as serviceWorker from './serviceWorker';
 import {BrowserRouter} from "react-router-dom";
 import {InMemoryCache} from "apollo-cache-inmemory";
@@ -17,7 +17,6 @@ const client = new ApolloClient({
         console.log('networkError', networkError)
     }
 });
-
 
 ReactDOM.render(
     <BrowserRouter>
